requirements (2).txt
torch>=2.0.0
transformers>=4.30.0
numpy>=1.22.0
pillow>=9.0.0
flask>=2.0.0
requests>=2.28.0
Werkzeug>=2.2.0
flask_cors>=3.0.10
bitsandbytes>=0.35.0
accelerate>=0.2.0
pyngrok>=6.0.0